refactor(datasets): Merge dataset services
