+++ web/src/features/mcp/features/datasets/tools/createDatasetRunItem.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { PostDatasetRunItemsV1Body } from "@/src/features/public-api/types/datasets";
-import { createDatasetRunItemForApi } from "@/src/features/public-api/server/dataset-run-items-api-service";
+import { createDatasetRunItemForApi } from "@/src/features/datasets/server/publicDatasetService";
 import { defineTool } from "../../../core/define-tool";
 import { runMcpTool } from "../../../core/run-mcp-tool";
 import { getMcpPublicApiAuth } from "../../publicApi";
